+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Release 5.6.5
+ * tbd
+
 # Release 5.6.4
  * Add: Koin Dependency Injection
  * Issuing: Fix regression not storing credentials in auth code flow